# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sing-box-agent API
  version: 1.2.0
  description: |
    A lightweight Go service that embeds sing-box as a library and provides a REST API for remote management.
    Designed to replace 3x-ui panels and SSH-based config management while maintaining multi-server architecture.

    **Key Features:**
    - Runtime user management without process restarts
    - Full protocol support via sing-box (VLESS, Hysteria2, ShadowTLS, TUIC, etc.)
    - Multi-server architecture — one agent per VPN server, controlled centrally
    - Prometheus metrics export for Grafana dashboards
    - Zero-dependency deployment (single binary)
    - State consistency — API is source of truth, agent applies desired state

  contact:
    name: sing-box-agent
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://agent.example.com
    description: Production server
security:
  - BearerAuth: []
  - RequestSigning: []
...
paths:
...
  /inbounds/{tag}/users:
    get:
      tags:
        - Users
      summary: List users
      description: Retrieve all users for a specific inbound
      operationId: listUsers
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
          description: Inbound tag identifier
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        '404':
          description: Inbound not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user for the specified inbound
      operationId: createUser
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
          description: Inbound tag identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserCreateResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inbounds/{tag}/users/{subId}:
    get:
      tags:
        - Users
      summary: Get user
      description: Retrieve a specific user by subscription ID
      operationId: getUser
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
          description: Inbound tag identifier
        - name: subId
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID (primary identifier)
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: USER_NOT_FOUND
                  message: User 'device-123' not found
    put:
      tags:
        - Users
      summary: Update user
      description: Update an existing user configuration
      operationId: updateUser
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
          description: Inbound tag identifier
        - name: subId
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user from the inbound
      operationId: deleteUser
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
          description: Inbound tag identifier
        - name: subId
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for authentication (minimum 32 characters)
    RequestSigning:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC-SHA256 signature for request authentication

        **Required Headers:**
        - X-Signature: HMAC-SHA256(canonical_string, secret)
        - X-Timestamp: Unix seconds (reject if |server_time - timestamp| > 300s)
        - X-Nonce: UUID v4 (stored in LRU cache, 24h TTL, 10k max entries)

        **Canonicalization:**
        ```
        canonical_string = "{nonce}\\n{timestamp}\\n{method}\\n{path}\\n{body_hash}"
        body_hash = SHA256(request_body).hexdigest()  // empty string if no body
        signature = HMAC-SHA256(canonical_string, secret).hexdigest()
        ```
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response payload
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - REPLAY_DETECTED
                - INBOUND_NOT_FOUND
                - USER_NOT_FOUND
                - USER_EXISTS
                - IDEMPOTENCY_CONFLICT
                - VERSION_CONFLICT
                - SYNC_FAILED
                - SERVICE_UNAVAILABLE
              description: Error code
            message:
              type: string
              description: Human-readable error message
...
    User:
      type: object
      required:
        - subId
        - uuid
      properties:
        subId:
          type: string
          description: Subscription ID (primary identifier)
        uuid:
          type: string
          format: uuid
          description: VLESS/VMess UUID (protocol-specific)
        name:
          type: string
          description: User display name
        email:
          type: string
          description: User email
        flow:
          type: string
          description: VLESS flow (e.g., "xtls-rprx-vision")
        enabled:
          type: boolean
          default: true
        traffic:
          type: object
          properties:
            upload:
              type: integer
              description: Cumulative upload bytes
            download:
              type: integer
              description: Cumulative download bytes
            reset_at:
              type: string
              format: date-time
              description: Last counter reset timestamp
    UserCreate:
      type: object
      required:
        - subId
        - uuid
      properties:
        subId:
          type: string
          description: Subscription ID
        uuid:
          type: string
          format: uuid
          description: VLESS/VMess UUID
        name:
          type: string
        email:
          type: string
        flow:
          type: string
        enabled:
          type: boolean
          default: true
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        flow:
          type: string
        enabled:
          type: boolean
    UserCreateResponse:
      type: object
      required:
        - subId
        - link
      properties:
        subId:
          type: string
          description: Subscription ID
        link:
          type: string
          description: Generated connection link
...